requirements.txt
yt_dlp>=2022.10.4
lyricsgenius>=3.0.1
eyed3==0.9.7
requests>=2.28.1
Pillow>=9.2.0